// src/Entities/Models/IUser.ts
import {ERoles} from "../Enums/ERoles";
import {ObjectId} from "mongodb";

export interface IUser {
    UserUniqeIdentifier : string
    _id?: ObjectId;

    // Personal
    firstName?: string;
    lastName?: string;
    email?: string;
    phone?: string;

    // Account
    username: string;
    role?: ERoles;
    createdAt?: string;
    loginAttempts?: number;
    updatedAt?: string;
    lockedUntil?: string;
    lastActiveAt?: string;
    isDisabled?: boolean;


    password?: string;

    // Set to true to disable main password, and only accept restore password token.
    // after this token is passed in request prompt for a new password to override
    // main password with.
    restorePassword?: boolean;
    // password generated by backend, send to user to restore password.
    restorePasswordToken?: string;
}